[Feature]实现首页点击菜品图片显示菜品详细界面的功能

+++ DianXiaoEr-Menu-iOS/Views/DXEOrderDishTableViewCell.m
@@ -26,7 +26,7 @@
 
 - (void)updateCellByDishCount:(NSInteger)count dishPrice:(float)price;
 {
-    self.dishCount.text = [NSString stringWithFormat:@"%ld", count];
+    self.dishCount.text = [NSString stringWithFormat:@"%d", (int)count];
     self.dishTotalPrice.text = [NSString stringWithFormat:@"￥%.2f", count * price];
     
     if (count == kDXEDishItemCountInCartMin)
